Fixed logged user delete bug.

+++ routes/user/delete_user.js
@@ -3,6 +3,7 @@ var router = express.Router();
 
 /* POST delete user. */
 router.post('/', isAuthenticated, function (req, res, next) {
+    if (!req.user.Id == req.body.UserId) {
         var delete_user_query = 'DELETE FROM User WHERE Id=\'' + req.body.UserId + '\'';
 
         database.query(delete_user_query, function (err, rows) {
@@ -10,6 +11,11 @@ router.post('/', isAuthenticated, function (req, res, next) {
             req.flash('info', 'Utente cancellato');
             res.redirect('/config/configuration');
         });
+    }
+    else {
+        req.flash('info', 'Impossibile cancellare l\'utente loggato');
+        res.redirect('/config/configuration');
+    }
     }
 );
